backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.5
pydantic_core==2.41.5


# Environment configuration
python-dotenv==1.2.1

# Type annotations and validation
typing_extensions==4.15.0
annotated-types==0.7.0

# Async support
anyio==4.12.0
sniffio==1.3.1

# HTTP utilities
h11==0.16.0
click==8.3.1

